+++ Render/MapRenderer.js
@@ -23,11 +23,17 @@ class MapRenderer {
     draw(renderContext, dungeonMap, mouseX, mouseY) {
         if (!renderContext) return
         if (!renderContext.showMap) return;
-        if (renderContext.hideInBoss && (Player.getX() > 0 || Player.getZ() > 0)) return;
+        if (renderContext.hideInBoss && this.tabs[1].shouldShowTab(renderContext, dungeonMap)) return;
 
         let { x, y, size } = renderContext.getMapDimensions()
 
-        if (renderContext.showTabs) this.drawTabs(renderContext, dungeonMap, mouseX, mouseY)
+        if (renderContext.showTabs) {
+            this.drawTabs(renderContext, dungeonMap, mouseX, mouseY)
+        } else {
+            for (let tab of this.tabs) {
+                tab.getRenderHeight(renderContext, dungeonMap) //Update wether to show the tab or not (bad coding sideeffects moment D:)
+            }
+        }
 
         // BACKROUND COLOR
         Renderer.drawRect(Renderer.color(renderContext.settings.mapBackgroundColor[0] ?? 0, renderContext.settings.mapBackgroundColor[1] ?? 0, renderContext.settings.mapBackgroundColor[2] ?? 0, renderContext.settings.mapBackgroundColor[3] ?? 150), x, y, size, size)//background
@@ -35,9 +41,9 @@ class MapRenderer {
         this.tabs[this.selectedTabIndex].draw(renderContext, dungeonMap, mouseX, mouseY)
 
         // THIS IS THE LEFT TOP AND RIGHT BORDER IN THAT ORDER
-        Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x, y, size, renderContext.borderWidth) //border
-        Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x, y, renderContext.borderWidth, size)
-        Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x + size - renderContext.borderWidth, y, renderContext.borderWidth, size)
+        Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x, y, size, renderContext.borderWidth) //border
+        Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x, y, renderContext.borderWidth, size)
+        Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x + size - renderContext.borderWidth, y, renderContext.borderWidth, size)
 
         // Dont render bottom line if scoreinfo rendering
         // Renderer.drawRect(Renderer.color(0, 0, 0), x, y + size - this.borderWidth, size, this.borderWidth)
@@ -65,10 +71,15 @@ class MapRenderer {
             let scoreInfoHeight = 10 * size / 100
             Renderer.drawRect(Renderer.color(renderContext.settings.extraInfoBackroundColor[0] ?? 0, renderContext.settings.extraInfoBackroundColor[1] ?? 0, renderContext.settings.extraInfoBackroundColor[2] ?? 0, renderContext.settings.extraInfoBackroundColor[3]), x, y + size, size, scoreInfoHeight)
 
-            renderLibs.drawStringCenteredFull("&f" + scoreInfo.total, x + size / 4, y + size + scoreInfoHeight / 2, size / 100)
+            renderLibs.drawStringCenteredFull("&f" + scoreInfo.total, x + size * (dungeonMap.floorNumber >= 6 ? 1 / 5 : 1 / 4), y + size + scoreInfoHeight / 2, size / 100)
+            renderLibs.drawStringCenteredFull((scoreInfo.crypts >= 5 ? `&a${scoreInfo.crypts}` : scoreInfo.crypts > 0 ? `&e${scoreInfo.crypts}` : `&c0`) + 'c', x + size * (dungeonMap.floorNumber >= 6 ? 10 / 23 : 3 / 4), y + size + scoreInfoHeight / 2, size / 100)
+            if (dungeonMap.floorNumber >= 6)
+                renderLibs.drawStringCenteredFull("&7Mimic " + (scoreInfo.mimic ? "&a✔" : "&c✕"), x + size / 4 * 3, y + size + scoreInfoHeight / 2, size / 100)
 
-            renderLibs.drawStringCenteredFull("&7Mimic " + (scoreInfo.mimic ? "&a✔" : "&c✕"), x + size / 4 * 3, y + size + scoreInfoHeight / 2, size / 100)
 
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x, y + size, renderContext.borderWidth, scoreInfoHeight) //border of score info
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x + size - renderContext.borderWidth, y + size, renderContext.borderWidth, scoreInfoHeight)
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x, y + size + scoreInfoHeight, size, renderContext.borderWidth)
 
             Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x, y + size, renderContext.borderWidth, scoreInfoHeight) //border of score info
             Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x + size - renderContext.borderWidth, y + size, renderContext.borderWidth, scoreInfoHeight)
@@ -97,9 +108,9 @@ class MapRenderer {
             renderLibs.drawStringCenteredShadow(mapLine1, x + size / 2, y + size + 1, size / 220)
             renderLibs.drawStringCenteredShadow(mapLine2, x + size / 2, y + size + 1 + 10 * size / 200, size / 220)
 
-            Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x, y + size, renderContext.borderWidth, scoreInfoHeight) //border of score info
-            Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x + size - renderContext.borderWidth, y + size, renderContext.borderWidth, scoreInfoHeight)
-            Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x, y + size + scoreInfoHeight, size, renderContext.borderWidth)
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x, y + size, renderContext.borderWidth, scoreInfoHeight) //border of score info
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x + size - renderContext.borderWidth, y + size, renderContext.borderWidth, scoreInfoHeight)
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x, y + size + scoreInfoHeight, size, renderContext.borderWidth)
 
             if (mouseX >= x && mouseX <= x + size
                 && mouseY >= y + size && mouseY <= y + size + scoreInfoHeight) {
@@ -108,7 +119,7 @@ class MapRenderer {
             }
         } else {
             // BOTTOM BORDER IF DISABLED
-            Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x, y + size, size, renderContext.borderWidth)
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x, y + size, size, renderContext.borderWidth)
         }
 
         if (renderContext.currentRoomInfo !== "none") {
@@ -149,9 +160,9 @@ class MapRenderer {
             renderLibs.scizzorFast(x + tabXOff, y - tabH, tabW, tabH)
 
             Renderer.drawRect(Renderer.color(renderContext.settings.mapBackgroundColor[0] ?? 0, renderContext.settings.mapBackgroundColor[1] ?? 0, renderContext.settings.mapBackgroundColor[2] ?? 0, renderContext.settings.mapBackgroundColor[3] ?? 150), x + tabXOff, y - maxTabH, tabW, maxTabH) //background
-            Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x + tabXOff, y - tabH, tabW, renderContext.borderWidth) //background
-            Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x + tabXOff, y - maxTabH, renderContext.borderWidth, maxTabH) //background
-            Renderer.drawRect(Renderer.color(renderContext.settings.borderColor[0] ?? 0, renderContext.settings.borderColor[1] ?? 0, renderContext.settings.borderColor[2] ?? 0, renderContext.settings.borderColor[3]), x + tabXOff + tabW - renderContext.borderWidth, y - maxTabH, renderContext.borderWidth, maxTabH) //background
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x + tabXOff, y - tabH, tabW, renderContext.borderWidth) //background
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x + tabXOff, y - maxTabH, renderContext.borderWidth, maxTabH) //background
+            Renderer.drawRect(Renderer.color(renderContext.settings.mapBorderColor[0] ?? 0, renderContext.settings.mapBorderColor[1] ?? 0, renderContext.settings.mapBorderColor[2] ?? 0, renderContext.settings.mapBorderColor[3]), x + tabXOff + tabW - renderContext.borderWidth, y - maxTabH, renderContext.borderWidth, maxTabH) //background
 
             let hovered = (mouseX >= x + tabXOff && mouseX <= x + tabXOff + tabW
                 && mouseY >= y - maxTabH && mouseY <= y)