+++ Render/RenderContext.js
@@ -9,7 +9,7 @@
  * @property {Number} iconScale - Width/height of icons (scales with size, will be same if size is 100)
  * @property {"hypixel"|"default"|"tenios"|"secrets"|"secrets_underhead"} tickStyle - Style of the ticks
  * @property {"none"|"text"|"icon"} puzzleNames - Render style of puzzle names
- * @property {Boolean} headBorder - Wether to put a black border around heads on the map
+ * @property {"none"|"single"|"class-color"|} headBorder - Wether to put a black border around heads on the map
  * @property {"never"|"leap"|"always"} playerNames - When to show player names on map
  * @property {"none"|"left"|"right"} currentRoomInfo - Render current room hover info on side of map
  * @property {"none"|"legalmap"|"simplified"} scoreInfoUnderMap - Render score info under the map
@@ -20,12 +20,13 @@
  * @property {Boolean} hideInBoss - Hide the map in boss entirely 
  * @property {Boolean} showTabs - Show tabs at the top of the map 
  * @property {Boolean} showSecrets - Show waypoints for secrets in the dungeon
+ * @property {Boolean} boxDoors - Put a box around wither doors
  * @property {Boolean} spiritLeapOverlay - Show an overlay on the spirit leap gui
  * @property {Boolean} forcePaul - Wether to force enable the +10 score for paul (eg if jerry mayor)
  * @property {Boolean} clearedRoomInfo - Show a summory of what rooms people cleared after run finishes
  * @property {String} apiKey - The user's api key, or "" if unknown
  * @property {Boolean} devInfo - Wether to show def info in various places in the map
- * @property {Object[]} borderColor - The RGBO value of the map border color
+ * @property {Object[]} mapBorderColor - The RGBO value of the map border color
  * @property {Object[]} mapBackgroundColor - The RGBO value of the map backround color
  * @property {Object[]} extraInfoBackroundColor - The RGBO value of the extrainfo backround color
  */
@@ -55,7 +56,7 @@ class RenderContext {
         return this.paddingLeft * 2 + this.blockSize * 6 + this.roomGap
     }
 
-    get showMap(){
+    get showMap() {
         return this.settings.showMap;
     }
 
@@ -139,6 +140,10 @@ class RenderContext {
 
     get clearedRoomInfo() {
         return this.settings.clearedRoomInfo
+    }
+
+    get boxDoors() {
+        return this.settings.boxDoors
     }
 
     get apiKey() {
@@ -258,7 +263,7 @@ class RenderContext {
         iconScale = 10,
         tickStyle = "default",
         puzzleNames = "none",
-        headBorder = false,
+        headBorder = "none",
         playerNames = "leap",
         currentRoomInfo = "none",
         scoreInfoUnderMap = "simplified",
@@ -274,9 +279,17 @@ class RenderContext {
         clearedRoomInfo = true,
         apiKey = "",
         devInfo = false,
-        borderColor = [0, 0, 0, 255],
+        boxDoors = true,
+        mapBorderColor = [0, 0, 0, 255],
         mapBackgroundColor = [0, 0, 0, 100],
-        extraInfoBackroundColor = [0, 0, 0, 100]
+        extraInfoBackroundColor = [0, 0, 0, 100],
+        healerColor = [240, 70, 240, 255],
+        mageColor = [70, 210, 210, 255],
+        bersColor = [255, 0, 0, 255],
+        archColor = [30, 170, 50, 255],
+        tankColor = [150, 150, 150, 255],
+        singleBorderColor = [0, 0, 0, 255]
+
     }) {
         return {
             showMap,
@@ -304,9 +317,16 @@ class RenderContext {
             clearedRoomInfo,
             apiKey,
             devInfo,
-            borderColor,
+            boxDoors,
+            mapBorderColor,
             mapBackgroundColor,
-            extraInfoBackroundColor
+            extraInfoBackroundColor,
+            healerColor,
+            mageColor,
+            bersColor,
+            archColor,
+            tankColor,
+            singleBorderColor
         }
     }