+++ Extra/Settings/Changelog.md
@@ -54,3 +54,23 @@ First released version!
 + Fix settings not saving correctly on linux
 + Setting to put player head above secrets when on secrets tick style
 ```
+
+@ver v1.0.9
+```diff
++ Added option for custom border colors depending on the class of the player (thenose2003)
++ Show crypts in simplified map info, only show mimic in f6 and above
++ White tick blood when all mobs spawned
++ Fix deaths on legalmap score info style not showing
++ Actual color pickers instead of 3 sliders
++ Update failed puzzles instantly from tab list
++ Instantly checkmark blaze puzzle after all blazes are done
++ box around wither doors on br
+! fix secrets under head setting showing secrets over head with tenios style
+! fix secret tick style showing ?/? for ironman players
+! fix spirit leap menu not switching to boss tab in boss
+! fix not switching to boss/score tab if you have tabs disabled
+! fix mimic kill detection
+! fix boss room detection for only show map in boss setting
+! fix changelog sometimes having weird gaps
+! fix boss map images not showing on most floors
+```